+++ src/service/BanService.java
@@ -18,7 +18,7 @@ public interface BanService {
     /**
      * 단일 반 검색하기
      */
-    BanDTO selectOneBan(int classNum);
+    BanDTO selectOneBan(long classNum);
 
     /**
      * 반 수정하기
